backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
boto3==1.28.0
stripe==7.0.0
cffi==1.17.1
click==8.1.8
cloudinary==1.38.0
colorama==0.4.6
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
google-generativeai==0.3.2
h11==0.14.0
httpx==0.28.1
idna==3.10
motor==3.7.0
passlib==1.7.4
Pillow==10.2.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pymongo==4.11
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
requests==2.31.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
mongomock==4.1.2
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.28.1
